Rename event/team info command
